requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
boto3==1.26.137
botocore==1.29.137
celery==5.5.3
certifi==2023.5.7
cffi==2.0.0
charset-normalizer==3.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.8.1
django-cors-headers==4.0.0
django-filter==25.2
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dotenv==0.9.9
flake8==7.3.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
Markdown==3.9
MarkupSafe==3.0.3
mccabe==0.7.0
nulltype==2.3.1
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
plaid==0.1.7
plaid-python==37.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.21
pyflakes==3.4.0
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.3
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
setuptools==80.9.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==13.0.1
truelayer-signing==0.3.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.15
uuid==1.30
vine==5.1.0
wcwidth==0.2.14
